+++ kousa/lib/business-logic/room_block_logic.ex
@@ -1,10 +1,12 @@
 defmodule Kousa.BL.RoomBlock do
-  alias Beef.Schemas.Users
+  alias Beef.Users
+  alias Beef.Rooms
+  alias Beef.RoomBlocks
 
   def unban(user_id, user_id_to_unban) do
     with {:ok, id} <- Users.tuple_get_current_room_id(user_id),
-         true <- Kousa.Data.Room.is_owner(id, user_id) do
-      Kousa.Data.RoomBlock.unban(id, user_id_to_unban)
+         true <- Rooms.is_owner(id, user_id) do
+          RoomBlock.unban(id, user_id_to_unban)
     end
   end
 
@@ -12,8 +14,8 @@ defmodule Kousa.BL.RoomBlock do
           false | {:err | nil | list, nil | number | {:error, :not_found}}
   def get_blocked_users(user_id, offset) do
     with {:ok, id} <- Users.tuple_get_current_room_id(user_id),
-         true <- Kousa.Data.Room.is_owner(id, user_id) do
-      Kousa.Data.RoomBlock.get_blocked_users(id, offset)
+         true <- Rooms.is_owner(id, user_id) do
+          RoomBlocks.get_blocked_users(id, offset)
     end
   end
 end
