+++ kousa/lib/business-logic/room_logic.ex
@@ -5,14 +5,14 @@ defmodule Kousa.BL.Room do
   alias Kousa.Gen
   alias Kousa.Caster
   alias Kousa.VoiceServerUtils
-  # note the following three module aliases are on the chopping block!
+  alias Beef.Users
+  alias Beef.Rooms
+  # note the following 2 module aliases are on the chopping block!
   alias Kousa.Data.Follower
-  alias Kousa.Data.Room
   alias Kousa.Data.RoomPermission
-  alias Beef.Schemas.Users
 
   def set_auto_speaker(user_id, value) do
-    room = Room.get_room_by_creator_id(user_id)
+    room = Rooms.get_room_by_creator_id(user_id)
 
     if not is_nil(room) do
       RegUtils.lookup_and_cast(Kousa.Gen.RoomSession, room.id, {:set_auto_speaker, value})
@@ -22,7 +22,7 @@ defmodule Kousa.BL.Room do
   @spec make_room_public(any, any) :: nil | :ok
   def make_room_public(user_id, new_name) do
     # this needs to be refactored if a user can have multiple rooms
-    case Room.set_room_privacy_by_creator_id(user_id, false, new_name) do
+    case Rooms.set_room_privacy_by_creator_id(user_id, false, new_name) do
       {1, [room]} ->
         Gen.RoomSession.send_cast(
           room.id,
@@ -38,7 +38,7 @@ defmodule Kousa.BL.Room do
   @spec make_room_private(any, any) :: nil | :ok
   def make_room_private(user_id, new_name) do
     # this needs to be refactored if a user can have multiple rooms
-    case Room.set_room_privacy_by_creator_id(user_id, true, new_name) do
+    case Rooms.set_room_privacy_by_creator_id(user_id, true, new_name) do
       {1, [room]} ->
         Kousa.Gen.RoomSession.send_cast(
           room.id,
@@ -57,7 +57,7 @@ defmodule Kousa.BL.Room do
     if not is_nil(user.currentRoomId) and
          Follower.is_following_me(user_id, user_id_to_invite) do
       # @todo store room name in RoomSession to avoid db lookups
-      room = Room.get_room_by_id(user.currentRoomId)
+      room = Rooms.get_room_by_id(user.currentRoomId)
 
       if not is_nil(room) do
         Gen.RoomSession.send_cast(
@@ -77,7 +77,7 @@ defmodule Kousa.BL.Room do
 
   def block_from_room(user_id, user_id_to_block_from_room) do
     with {status, room} when status in [:creator, :mod] <-
-           Room.get_room_status(user_id) do
+           Rooms.get_room_status(user_id) do
       if room.creatorId != user_id_to_block_from_room do
         RoomBlock.insert(%{
           modId: user_id,
@@ -111,7 +111,7 @@ defmodule Kousa.BL.Room do
         Users.get_current_room_id(user_id_to_set_listener)
       )
     else
-      {status, room} = Room.get_room_status(user_id)
+      {status, room} = Rooms.get_room_status(user_id)
       is_creator = user_id_to_set_listener == not is_nil(room) and room.creatorId
 
       if not is_creator and (status == :creator or status == :mod) do
@@ -146,13 +146,13 @@ defmodule Kousa.BL.Room do
 
   def make_speaker(user_id, user_id_to_make_speaker) do
     with {status, room} when status in [:creator, :mod] <-
-           Room.get_room_status(user_id) do
+           Rooms.get_room_status(user_id) do
       internal_set_speaker(user_id_to_make_speaker, room.id)
     end
   end
 
   def change_mod(user_id, user_id_to_change, value) do
-    room = Room.get_room_by_creator_id(user_id)
+    room = Rooms.get_room_by_creator_id(user_id)
 
     if room do
       RoomPermission.set_is_mod(user_id_to_change, room.id, Caster.bool(value))
@@ -197,7 +197,7 @@ defmodule Kousa.BL.Room do
   # @deprecated
   def rename_room(user_id, new_name) do
     with {:ok, room_id} <- Users.tuple_get_current_room_id(user_id),
-         {1, _} <- Room.update_name(user_id, new_name) do
+         {1, _} <- Rooms.update_name(user_id, new_name) do
       nil
       RegUtils.lookup_and_cast(Gen.RoomSession, room_id, {:new_room_name, new_name})
     end
@@ -205,8 +205,8 @@ defmodule Kousa.BL.Room do
 
   def edit_room(user_id, new_name, new_description, is_private) do
 
-    with {:ok, room_id} <- Data.User.tuple_get_current_room_id(user_id) do
-      case Data.Room.edit(room_id, %{name: new_name, description: new_description, is_private: is_private}) do
+    with {:ok, room_id} <- Users.tuple_get_current_room_id(user_id) do
+      case Rooms.edit(room_id, %{name: new_name, description: new_description, is_private: is_private}) do
 
         {:ok, _room} ->
           RegUtils.lookup_and_cast(Gen.RoomSession, room_id, {:new_room_details, new_name, new_description, is_private})
@@ -229,7 +229,7 @@ defmodule Kousa.BL.Room do
 
     id = Ecto.UUID.generate()
 
-    case Room.create(%{
+    case Rooms.create(%{
            id: id,
            name: room_name,
            description: room_description,
@@ -286,9 +286,9 @@ defmodule Kousa.BL.Room do
     currentRoomId = Users.get_current_room_id(user_id)
 
     if currentRoomId == room_id do
-      %{room: Room.get_room_by_id(room_id)}
+      %{room: Rooms.get_room_by_id(room_id)}
     else
-      case Room.can_join_room(room_id, user_id) do
+      case Rooms.can_join_room(room_id, user_id) do
         {:error, message} ->
           %{error: message}
 
@@ -317,7 +317,7 @@ defmodule Kousa.BL.Room do
                 leave_room(user_id, currentRoomId)
               end
 
-              updated_user = Room.join_room(room, user_id)
+              updated_user = Rooms.join_room(room, user_id)
 
               Kousa.Gen.RoomSession.send_cast(
                 room_id,
@@ -347,7 +347,7 @@ defmodule Kousa.BL.Room do
         else: current_room_id
 
     if current_room_id do
-      case Room.leave_room(user_id, current_room_id) do
+      case Rooms.leave_room(user_id, current_room_id) do
         # the room should be destroyed
         {:bye, room} ->
           Kousa.Gen.RoomSession.send_cast(current_room_id, {:destroy, user_id})
