+++ kousa/lib/business-logic/user_logic.ex
@@ -1,9 +1,8 @@
 defmodule Kousa.BL.User do
   alias Kousa.Gen
   alias Kousa.RegUtils
-  alias Beef.Schemas.Users
+  alias Beef.Users
   alias Kousa.Gen
-  alias Kousa.Data
   alias Kousa.RegUtils
   alias Kousa.BL
