+++ kousa/lib/socket_handler.ex
@@ -5,7 +5,8 @@ defmodule Kousa.SocketHandler do
   alias Kousa.RegUtils
   alias Kousa.Gen
   alias Kousa.Caster
-  alias Beef.Schemas.Users
+  alias Beef.Users
+  alias Beef.Rooms
 
   # TODO: just collapse this into its parent module.
   defmodule State do
@@ -143,7 +144,7 @@ defmodule Kousa.SocketHandler do
                     cond do
                       not is_nil(user.currentRoomId) ->
                         # @todo this should probably go inside room business logic
-                        room = Kousa.Data.Room.get_room_by_id(user.currentRoomId)
+                        room = Rooms.get_room_by_id(user.currentRoomId)
 
                         {:ok, room_session} =
                           GenRegistry.lookup_or_start(Gen.RoomSession, user.currentRoomId, [
@@ -560,7 +561,7 @@ defmodule Kousa.SocketHandler do
 
   def handler("ask_to_speak", _data, state) do
     with {:ok, room_id} <- Users.tuple_get_current_room_id(state.user_id) do
-      case Kousa.Data.RoomPermission.ask_to_speak(state.user_id, room_id) do
+      case RoomsPermission.ask_to_speak(state.user_id, room_id) do
         {:ok, %{isSpeaker: true}} ->
           Kousa.BL.Room.internal_set_speaker(state.user_id, room_id)
 
@@ -639,7 +640,7 @@ defmodule Kousa.SocketHandler do
 
   def f_handler("get_top_public_rooms", data, %State{} = state) do
     {rooms, next_cursor} =
-      Kousa.Data.Room.get_top_public_rooms(
+      Rooms.get_top_public_rooms(
         state.user_id,
         data["cursor"]
       )
