+++ kousa/test/_support/factory.ex
@@ -10,8 +10,8 @@ defmodule Kousa.Support.Factory do
 
   alias Beef.Repo
   alias Beef.Schemas.User
-  alias Beef.Room
   alias Beef.Schemas.ScheduledRoom
+  alias Beef.Schemas.Room
 
   def create(struct, data \\ [])
